Socketry::Resolver: pluggable DNS resolution support with timeouts
This provides an abstract API for mapping IP addresses and domain names to
Resolv::IPv4 and Resolv::IPv6 addresses.

It provides two backends:

* Socketry::Resolver::System: based on IPSocket.getaddress, which is implemented
  using the POSIX getaddrinfo(3) function, with timeouts provided by Ruby's
  Timeout module.
* Socketry::Resolver::Resolv: based on Resolv::DNS, a pure Ruby implementation
  of the DNS protocol, which natively implements timeouts using Ruby's
  asynchronous I/O primitives.

I have made Socketry::Resolver::System the default. I am aware that Ruby's
Timeout is broken in a multithreaded context[1] (that is, of course, the whole
point of this library existing in the first place), but I think it will be a
generally less problematic default, as I have experienced a long history of
issues with non-system resolvers.